refactor: 삭제 시 요소가 존재하지 않을 경우 400 에러를 반환하도록 수정

+++ src/main/java/roomescape/service/ReservationService.java
@@ -2,6 +2,7 @@ package roomescape.service;
 
 import java.time.LocalDate;
 import java.util.List;
+import java.util.NoSuchElementException;
 import org.springframework.stereotype.Service;
 import roomescape.domain.Reservation;
 import roomescape.domain.ReservationTime;
@@ -81,6 +82,9 @@ public class ReservationService {
     }
 
     public ReservationDeleteResponse delete(final long id) {
+        if (reservationDao.findById(id).isEmpty()) {
+            throw new NoSuchElementException();
+        }
         return new ReservationDeleteResponse(reservationDao.delete(id));
     }
 }
